12/17/17
Added and tested the GetMenuPick function and also added a template to allow the to_string function to work with C++98. Menu items are currently commeneted out for testing purposes and unsiged int data types where used for the iterator i value in the for loops that use the string.length function.